@@ -1,8 +1,9 @@
 {
   "name": "functions",
   "scripts": {
-    "lint": "tslint --project tsconfig.json",
+    "lint": "eslint --ext .js,.ts .",
     "build": "tsc",
+    "build:watch": "tsc --watch",
     "serve": "npm run build && firebase emulators:start --only functions",
     "shell": "npm run build && firebase functions:shell",
     "start": "npm run shell",
@@ -11,29 +12,30 @@
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "8"
+    "node": "16"
   },
   "main": "lib/index.js",
   "dependencies": {
-    "axios": "^0.21.1",
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "firebase-admin": "^8.10.0",
-    "firebase-functions": "^3.6.1"
+    "firebase-admin": "^10.0.2",
+    "firebase-functions": "^3.18.0"
   },
   "devDependencies": {
-    "@types/chai": "^4.2.11",
-    "@types/cors": "^2.8.6",
-    "@types/express": "^4.17.6",
-    "@types/mocha": "^7.0.2",
-    "@types/sinon": "^9.0.0",
-    "chai": "^4.2.0",
-    "firebase-functions-test": "^0.2.1",
-    "mocha": "^7.1.2",
-    "sinon": "^9.0.2",
-    "ts-node": "^8.9.1",
-    "tslint": "^5.12.0",
-    "typescript": "^3.8.0"
+    "@types/chai": "^4.3.4",
+    "@types/cors": "^2.8.13",
+    "@types/express": "^4.17.15",
+    "@types/mocha": "^10.0.1",
+    "@types/sinon": "^10.0.13",
+    "@typescript-eslint/eslint-plugin": "^5.12.0",
+    "@typescript-eslint/parser": "^5.12.0",
+    "chai": "^4.3.7",
+    "eslint": "^8.9.0",
+    "eslint-config-google": "^0.14.0",
+    "eslint-plugin-import": "^2.25.4",
+    "firebase-functions-test": "^0.2.0",
+    "mocha": "^10.2.0",
+    "sinon": "^15.0.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.5.4"
   },
   "private": true
 }
