@@ -1,4 +1,4 @@
-import 'types';
+import "types";
 
 declare module "solver" {
   export function solve(preferences: PreferenceRecord): Assignment;
